yaw控制加入前馈，但是pid不行

+++ BSP/Inc/bsp_uart.h
@@ -49,7 +49,7 @@ typedef struct {
 extern gambleState gamble_state;
 extern int32_t pushrot_M2006_positionTarget;
 extern int16_t pushrot_M2006_speedTarget;
-extern int32_t Load_M3508_positionTarget;
+extern float Load_M3508_positionTarget;
 extern int16_t Load_M3508_speedTarget;
 extern int16_t M3508Friction_speedTarget[3];
 extern float YAW_D4310_positiontarget;
